slyd/requirements.txt
# Slyd requirements
twisted==16.1.1
service_identity==16.0.0
requests==2.9.1
autobahn==0.13.0
six==1.10.0
chardet==2.3.0
monotonic==0.3
marshmallow_jsonapi==0.7.0
parse==1.6.6
ndg-httpsclient==0.4.0
django==1.9.4
mock==2.0.0

# Dulwich's MySQL backend
mysql-python==1.2.5
dulwich==0.13.0

# Splash dev
--allow-external=splash
https://github.com/scrapinghub/splash/archive/master.tar.gz
